fix npm publish to include dist files by flattening ./dist, clean on build, move all config options to tsconfig.json
