+++ .npmignore
@@ -3,4 +3,4 @@ npm-debug.log
 .vscode
 
 tests
-src
+/src
