+++ src/tests/Toxic.ts
@@ -1,6 +1,6 @@
 import { test } from "ava";
-import { createProxy, createToxic } from "../src/TestHelper";
-import { Latency } from "../src/Toxic";
+import { createProxy, createToxic } from "../TestHelper";
+import { Latency } from "../Toxic";
 
 test("Toxic Should remove a toxic", async (t) => {
   const { proxy } = await createProxy(t, "remove-toxic-test");
